+++ packages/pluggableWidgets/datagrid-web/typings/global.d.ts
@@ -1,4 +1,4 @@
-import { DATAGRID_DATA_EXPORT, DataExporter, DataGridName } from "../src/features/export";
+import { DATAGRID_DATA_EXPORT, DataExporter, DataGridName } from "../src/features/data-export"
 
 declare global {
     interface Window {
